replaced players and playersheet contexts with hooks
