+++ src/components/players/players-table.tsx
@@ -1,4 +1,8 @@
-import type { Player } from "@/lib/types";
+import type {
+  Player,
+  UsePlayersReturn,
+  UsePlayersSheetReturn,
+} from "@/lib/types";
 import { flexRender } from "@tanstack/react-table";
 import type { Table as TanstackTable } from "@tanstack/react-table";
 import {
@@ -14,9 +18,16 @@ import { ActionContextMenu } from "./action-context-menu";
 type PlayersTableProps = {
   table: TanstackTable<Player>;
   numColumns: number;
+  deletePlayer: UsePlayersReturn["deletePlayer"];
+  openPlayerSheet: UsePlayersSheetReturn["openPlayerSheet"];
 };
 
-export function PlayersTable({ table, numColumns }: PlayersTableProps) {
+export function PlayersTable({
+  table,
+  numColumns,
+  deletePlayer,
+  openPlayerSheet,
+}: PlayersTableProps) {
   return (
     <div className="border">
       <Table>
@@ -41,11 +52,13 @@ export function PlayersTable({ table, numColumns }: PlayersTableProps) {
         <TableBody>
           {table.getRowModel().rows?.length ? (
             table.getRowModel().rows.map((row) => (
-              <ActionContextMenu key={row.id} id={row.original.id}>
-                <TableRow
-                  key={row.id}
-                  data-state={row.getIsSelected() && "selected"}
-                >
+              <ActionContextMenu
+                key={row.id}
+                id={row.original.id}
+                deletePlayer={deletePlayer}
+                openPlayerSheet={openPlayerSheet}
+              >
+                <TableRow data-state={row.getIsSelected() && "selected"}>
                   {row.getVisibleCells().map((cell) => (
                     <TableCell key={cell.id}>
                       {flexRender(
