--- src/contexts/player-sheet-context.tsx
@@ -1,6 +1,0 @@
-import { createContext } from "react";
-import type { PlayerSheetContextType } from "@/lib/types.ts";
-
-export const PlayerSheetContext = createContext<
-  PlayerSheetContextType | undefined
->(undefined);
