+++ src/main.tsx
@@ -5,14 +5,11 @@ import { RouterProvider, createRouter } from "@tanstack/react-router";
 import { routeTree } from "./routeTree.gen";
 import { useAuth } from "./hooks/useAuth";
 import { AuthProvider } from "./contexts/auth-provider";
-import { usePlayers } from "./hooks/usePlayers";
-import { PlayersProvider } from "./contexts/players-provider";
 
 const router = createRouter({
   routeTree,
   context: {
     auth: undefined!,
-    players: undefined!,
   },
 });
 
@@ -24,16 +21,13 @@ declare module "@tanstack/react-router" {
 
 function InnerApp() {
   const auth = useAuth();
-  const players = usePlayers();
-  return <RouterProvider router={router} context={{ auth, players }} />;
+  return <RouterProvider router={router} context={{ auth }} />;
 }
 
 function App() {
   return (
     <AuthProvider>
-      <PlayersProvider>
-        <InnerApp />
-      </PlayersProvider>
+      <InnerApp />
     </AuthProvider>
   );
 }
