+++ src/routes/__root.tsx
@@ -1,10 +1,9 @@
-import type { AuthContextType, PlayersContextType } from "@/lib/types";
+import type { AuthContextType } from "@/lib/types";
 import { createRootRouteWithContext, Outlet } from "@tanstack/react-router";
 import { TanStackRouterDevtools } from "@tanstack/react-router-devtools";
 
 type RouterContext = {
   auth: AuthContextType;
-  players: PlayersContextType;
 };
 
 export const Route = createRootRouteWithContext<RouterContext>()({
